Satisfy theme PR guardrails before review rerun
The PR was blocked by Swift file-length budget and review findings around tagged-build theme config ownership. The fix keeps the theme behavior intact while splitting the new settings/UI/test code into smaller files, threading the active bundle identifier through CLI managed-config writes, and making the adaptive picker option always reachable.

Constraint: workflow-guard-tests rejects new over-budget Swift file growth

Constraint: tagged debug/staging bundles need managed theme writes to use the active bundle identifier

Rejected: Refresh the Swift file-length budget | the guard output requested splitting or reducing growth instead of accepting new debt

Confidence: high

Scope-risk: moderate

Tested: ./scripts/reload.sh --tag issue-3523-theme-system-separation

Tested: python3 scripts/swift_file_length_budget.py --budget .github/swift-file-length-budget.tsv

Tested: python3 -m json.tool Resources/Localizable.xcstrings

Not-tested: Local XCTest run; project policy says tests run in CI
