+++ Sources/TerminalThemeSettings.swift
@@ -1,40 +1,6 @@
 import Darwin
 import Foundation
 
-enum TerminalThemeMode: Equatable, Hashable {
-    case custom
-    case named(String)
-    case adaptive(light: String, dark: String)
-
-    var rawThemeValue: String? {
-        switch self {
-        case .custom:
-            return nil
-        case .named(let name):
-            let trimmed = name.trimmingCharacters(in: .whitespacesAndNewlines)
-            return trimmed.isEmpty ? nil : trimmed
-        case .adaptive(let light, let dark):
-            return TerminalThemeSettings.encodedThemeValue(light: light, dark: dark)
-        }
-    }
-}
-
-struct TerminalThemeSelection: Equatable {
-    let mode: TerminalThemeMode
-    let rawValue: String?
-    let light: String?
-    let dark: String?
-    let sourcePath: String?
-
-    static let custom = TerminalThemeSelection(
-        mode: .custom,
-        rawValue: nil,
-        light: nil,
-        dark: nil,
-        sourcePath: nil
-    )
-}
-
 enum TerminalThemeSettings {
     static let defaultManagedBundleIdentifier = "com.cmuxterm.app"
     static let managedBlockStart = "# cmux themes start"
@@ -226,6 +192,7 @@ enum TerminalThemeSettings {
             for: .applicationSupportDirectory,
             in: .userDomainMask
         ).first,
+        bundleIdentifier: String = defaultManagedBundleIdentifier,
         fileManager: FileManager = .default,
         reload: () -> Void = {}
     ) throws -> URL {
@@ -238,11 +205,13 @@ enum TerminalThemeSettings {
             configURL = try writeManagedThemeOverride(
                 rawThemeValue: rawThemeValue,
                 appSupportDirectory: appSupportDirectory,
+                bundleIdentifier: bundleIdentifier,
                 fileManager: fileManager
             )
         } else {
             configURL = try clearManagedThemeOverride(
                 appSupportDirectory: appSupportDirectory,
+                bundleIdentifier: bundleIdentifier,
                 fileManager: fileManager
             )
         }
@@ -274,92 +243,6 @@ enum TerminalThemeSettings {
         return lastValue
     }
 
-    static func parseSelection(rawValue: String?, sourcePath: String?) -> TerminalThemeSelection {
-        guard let rawValue = rawValue?.trimmingCharacters(in: .whitespacesAndNewlines), !rawValue.isEmpty else {
-            return TerminalThemeSelection(
-                mode: .custom,
-                rawValue: nil,
-                light: nil,
-                dark: nil,
-                sourcePath: sourcePath
-            )
-        }
-
-        var fallbackTheme: String?
-        var lightTheme: String?
-        var darkTheme: String?
-
-        for token in rawValue.split(separator: ",").map(String.init) {
-            let entry = token.trimmingCharacters(in: .whitespacesAndNewlines)
-            guard !entry.isEmpty else { continue }
-
-            let parts = entry.split(separator: ":", maxSplits: 1).map(String.init)
-            if parts.count != 2 {
-                if fallbackTheme == nil {
-                    fallbackTheme = entry
-                }
-                continue
-            }
-
-            let key = parts[0].trimmingCharacters(in: .whitespacesAndNewlines).lowercased()
-            let value = parts[1].trimmingCharacters(in: .whitespacesAndNewlines)
-            guard !value.isEmpty else { continue }
-
-            switch key {
-            case "light":
-                if lightTheme == nil {
-                    lightTheme = value
-                }
-            case "dark":
-                if darkTheme == nil {
-                    darkTheme = value
-                }
-            default:
-                if fallbackTheme == nil {
-                    fallbackTheme = value
-                }
-            }
-        }
-
-        let resolvedLight = lightTheme ?? fallbackTheme
-        let resolvedDark = darkTheme ?? fallbackTheme
-        let mode: TerminalThemeMode
-        if let lightTheme, let darkTheme, lightTheme.caseInsensitiveCompare(darkTheme) != .orderedSame {
-            mode = .adaptive(light: lightTheme, dark: darkTheme)
-        } else if let theme = resolvedDark ?? resolvedLight {
-            mode = .named(theme)
-        } else {
-            mode = .custom
-        }
-
-        return TerminalThemeSelection(
-            mode: mode,
-            rawValue: rawValue,
-            light: resolvedLight,
-            dark: resolvedDark,
-            sourcePath: sourcePath
-        )
-    }
-
-    static func encodedThemeValue(light: String?, dark: String?) -> String? {
-        let normalizedLight = light?.trimmingCharacters(in: .whitespacesAndNewlines)
-        let normalizedDark = dark?.trimmingCharacters(in: .whitespacesAndNewlines)
-
-        switch (normalizedLight?.isEmpty == false ? normalizedLight : nil, normalizedDark?.isEmpty == false ? normalizedDark : nil) {
-        case let (lightTheme?, darkTheme?):
-            if lightTheme.caseInsensitiveCompare(darkTheme) == .orderedSame {
-                return lightTheme
-            }
-            return "light:\(lightTheme),dark:\(darkTheme)"
-        case let (lightTheme?, nil):
-            return "light:\(lightTheme)"
-        case let (nil, darkTheme?):
-            return "dark:\(darkTheme)"
-        case (nil, nil):
-            return nil
-        }
-    }
-
     @discardableResult
     static func writeManagedThemeOverride(
         rawThemeValue: String,
@@ -367,17 +250,18 @@ enum TerminalThemeSettings {
             for: .applicationSupportDirectory,
             in: .userDomainMask
         ).first,
+        bundleIdentifier: String = defaultManagedBundleIdentifier,
         fileManager: FileManager = .default
     ) throws -> URL {
         guard let appSupportDirectory else {
             throw CocoaError(.fileNoSuchFile)
         }
-        let configURL = managedConfigURL(appSupportDirectory: appSupportDirectory)
+        let configURL = managedConfigURL(appSupportDirectory: appSupportDirectory, bundleIdentifier: bundleIdentifier)
         let directoryURL = configURL.deletingLastPathComponent()
         try fileManager.createDirectory(at: directoryURL, withIntermediateDirectories: true, attributes: nil)
 
         let existingContents = try readOptionalThemeOverrideContents(at: configURL) ?? ""
-        let strippedContents = removingManagedThemeOverride(from: existingContents)
+        let strippedContents = removingManagedThemeState(from: existingContents)
             .trimmingCharacters(in: .whitespacesAndNewlines)
         let block = """
         \(managedBlockStart)
@@ -396,17 +280,18 @@ enum TerminalThemeSettings {
             for: .applicationSupportDirectory,
             in: .userDomainMask
         ).first,
+        bundleIdentifier: String = defaultManagedBundleIdentifier,
         fileManager: FileManager = .default
     ) throws -> URL {
         guard let appSupportDirectory else {
             throw CocoaError(.fileNoSuchFile)
         }
-        let configURL = managedConfigURL(appSupportDirectory: appSupportDirectory)
+        let configURL = managedConfigURL(appSupportDirectory: appSupportDirectory, bundleIdentifier: bundleIdentifier)
         guard let existingContents = try readOptionalThemeOverrideContents(at: configURL) else {
             return configURL
         }
 
-        let strippedContents = removingManagedThemeOverride(from: existingContents)
+        let strippedContents = removingManagedThemeState(from: existingContents)
             .trimmingCharacters(in: .whitespacesAndNewlines)
 
         if strippedContents.isEmpty {
@@ -424,13 +309,15 @@ enum TerminalThemeSettings {
         return configURL
     }
 
-    static func removingManagedThemeOverride(from contents: String) -> String {
+    static func removingManagedThemeState(from contents: String) -> String {
         let pattern = #"(?ms)\n?# cmux themes start\n.*?\n# cmux themes end\n?"#
         guard let regex = try? NSRegularExpression(pattern: pattern) else {
             return contents
         }
         let fullRange = NSRange(contents.startIndex..<contents.endIndex, in: contents)
         let withoutManagedBlock = regex.stringByReplacingMatches(in: contents, options: [], range: fullRange, withTemplate: "")
+        // Older cmux builds wrote bare theme directives into this cmux-owned file.
+        // Custom mode should remove those legacy directives too.
         return withoutManagedBlock
             .components(separatedBy: .newlines)
             .filter { !isThemeDirectiveLine($0) }
